Add type hints

+++ src/Y15/D03.hs
@@ -21,7 +21,7 @@ solve1 :: String -> Int
 solve1 = length . nub . moves2houses
 
 solve2 :: String -> Int
-solve2 xs = let (santaPairs, robotPairs) = partition (even . fst) $ zip [0..] xs
+solve2 xs = let (santaPairs, robotPairs) = partition (even . fst) $ zip [(0::Int)..] xs
                 santaHouses = moves2houses $ map snd santaPairs
                 robotHouses = moves2houses $ map snd robotPairs
              in length . nub $ santaHouses ++ robotHouses
